refactor(spa-editor): build workout-store selector registry (§4.4)
Split the 115-LOC `workout-store-selectors.ts` into seven domain-grouped
files under `store/selectors/` (workout, step, selection, repetition
block, history, modal) and moved the two composite hooks
(`useKeyboardStoreSelectors`, `useContextMenuStore`) into the same
registry. All files now sit at <=61 LOC. A single barrel
`store/selectors/index.ts` re-exports the full surface, and every
callsite (10 source files + 5 test mocks) imports from the barrel
instead of the individual modules.
