+++ packages/workout-spa-editor/src/store/selectors/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Selector registry — single entry point for all workout-store
+ * selector hooks (atomic + composite). Callers should import from
+ * `store/selectors` rather than reaching into individual files so
+ * grouping changes stay an internal refactor.
+ */
+
+export * from "./history-selectors";
+export * from "./modal-selectors";
+export * from "./repetition-block-selectors";
+export * from "./selection-selectors";
+export * from "./step-selectors";
+export { useContextMenuStore } from "./use-context-menu-store";
+export { useKeyboardStoreSelectors } from "./use-keyboard-store-selectors";
+export * from "./workout-selectors";
